Switched to official mongo driver

+++ event_store.go
@@ -1,8 +1,9 @@
 package goengine
 
+// EventStore ...
 type EventStore interface {
 	Append(events *EventStream) error
 	GetEventsFor(streamName StreamName, id string) (*EventStream, error)
 	FromVersion(streamName StreamName, id string, version int) (*EventStream, error)
-	CountEventsFor(streamName StreamName, id string) (int, error)
+	CountEventsFor(streamName StreamName, id string) (int64, error)
 }
